refactor(core): make core own lexer implementation
