+++ CLAUDE.md
@@ -37,13 +37,14 @@ Do not restate feature behavior in this file. Read the canonical source instead.
 
 ```
 crates/                           # 38 publishable crates (36 core + 2 facade)
-  Parser:     copybook-core, copybook-lexer
+  Parser:     copybook-core
   Codec:      copybook-codec, copybook-codec-memory, copybook-codepage,
               copybook-charset, copybook-overpunch, copybook-zoned-format
   CLI:        copybook-cli, copybook-cli-determinism, copybook-options
   Framing:    copybook-fixed, copybook-rdw, copybook-rdw-predicates,
               copybook-record-io
-  Schema:     copybook-dialect, copybook-determinism, copybook-support-matrix
+  Schema:     copybook-determinism, copybook-support-matrix
+  Compatibility: copybook-dialect, copybook-lexer
   Governance: copybook-contracts, copybook-governance,
               copybook-governance-contracts, copybook-governance-grid,
               copybook-governance-runtime
