+++ crates/claude.md
@@ -10,11 +10,12 @@
 | Layer | Crates |
 |---|---|
 | Facade | `copybook`, `copybook-rs` |
-| Parser | `copybook-core`, `copybook-lexer` |
+| Parser | `copybook-core` |
 | Codec | `copybook-codec`, `copybook-codec-memory`, `copybook-codepage`, `copybook-charset`, `copybook-overpunch`, `copybook-zoned-format` |
 | CLI | `copybook-cli`, `copybook-cli-determinism`, `copybook-options` |
 | Framing | `copybook-fixed`, `copybook-rdw`, `copybook-rdw-predicates`, `copybook-record-io` |
-| Schema | `copybook-dialect`, `copybook-determinism`, `copybook-support-matrix` |
+| Schema | `copybook-determinism`, `copybook-support-matrix` |
+| Compatibility | `copybook-dialect`, `copybook-lexer` |
 | Governance | `copybook-contracts`, `copybook-governance`, `copybook-governance-contracts`, `copybook-governance-grid`, `copybook-governance-runtime` |
 | Safety | `copybook-error`, `copybook-error-reporter`, `copybook-overflow`, `copybook-safe-index`, `copybook-safe-ops`, `copybook-safe-text`, `copybook-utils` |
 | Quality | `copybook-corruption`, `copybook-corruption-detectors`, `copybook-corruption-predicates`, `copybook-corruption-rdw` |
