+++ crates/copybook-lexer/Cargo.toml
@@ -20,10 +20,7 @@ all-features = false
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
-logos.workspace = true
-
-[dev-dependencies]
-proptest.workspace = true
+copybook-core.workspace = true
 
 [lints]
 workspace = true
