+++ fuzz/fuzz_targets/fuzz_lexer.rs
@@ -1,7 +1,7 @@
 #![no_main]
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-use copybook_lexer::{CobolFormat, Lexer, LexerOptions, Token};
+use copybook_core::lexer::{CobolFormat, Lexer, LexerOptions, Token};
 use libfuzzer_sys::fuzz_target;
 
 // Hardened fuzz target for the COBOL lexer.
