+++ fuzz/fuzz_targets/lexer.rs
@@ -1,7 +1,7 @@
 #![no_main]
 // SPDX-License-Identifier: AGPL-3.0-or-later
 use libfuzzer_sys::fuzz_target;
-use copybook_lexer::{Lexer, LexerOptions, Token};
+use copybook_core::lexer::{Lexer, LexerOptions, Token};
 
 /// Fuzz target for lexer tokenization.
 ///
